requirements.txt
aspose-pdf
mysql-connector-python
Pillow
matplotlib
pdfminer.six
pymysql
python-docx
reportlab
tkinter